Revert "Creates PasskeyVerifiedModal.jsx, VerifyOtherWaysModal.jsx, VerifyWithEmailModal.jsx, and implements them on UpdatePoliticianInformation.jsx"
This reverts commit d81558574e32599f13c8d5b85a768915e19d97f1.
